feat: ModuleCache is only exposed in wasmer_sys mode, build_module() is only exposed in wasmer_wamr mode, rename error variant for accuracy
